django_api/requirements/development.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2-binary==2.9.7
python-slugify==8.0.1
pytz==2023.3.post1
slugify==0.0.1
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
